requirements.txt
celery
fastapi
keras-bert
keras-metrics
numpy
pandas
redis
sklearn
tensorflow_addons
tqdm
uvicorn
